+++ lib/src/main/java/com/wjduquette/joe/StandardLibrary.java
@@ -19,7 +19,7 @@ public class StandardLibrary extends Library {
         globalFunction("stringify", this::_stringify);
         globalFunction("typeName",  this::_typeName);
 
-        type(DoubleProxy.TYPE);
+        type(NumberProxy.TYPE);
         type(ErrorProxy.TYPE);
         type(KeywordProxy.TYPE);
         type(PairProxy.TYPE);