+++ lib/src/main/java/com/wjduquette/joe/types/NumberProxy.java
@@ -5,14 +5,14 @@ import com.wjduquette.joe.TypeProxy;
 
 import java.util.List;
 
-public class DoubleProxy extends TypeProxy<Double> {
-    public static final DoubleProxy TYPE = new DoubleProxy();
+public class NumberProxy extends TypeProxy<Double> {
+    public static final NumberProxy TYPE = new NumberProxy();
 
     //-------------------------------------------------------------------------
     // Constructor
 
-    public DoubleProxy() {
-        super("Double");
+    public NumberProxy() {
+        super("Number");
         staticType();
 
         constant("PI", Math.PI);
@@ -24,7 +24,7 @@ public class DoubleProxy extends TypeProxy<Double> {
     // Static Methods
 
     private Object _abs(Joe joe, List<Object> args) {
-        Joe.exactArity(args, 1, "Double.abs(value)");
+        Joe.exactArity(args, 1, "Number.abs(value)");
         return Math.abs(joe.toDouble(args.get(0)));
     }
 }